Fixed audio bug where music was playing simultaniously
Fixed audio bug where music was playing simultaniously.
The bug occurred when the user navigates before the music track had finished decoding in the browser and we then try to fade that track out.
The fix was to stop the music completely rather than fade out if the track was still decoding.