CGPROD-2124 Dirty fix for progress bar progress
Progress bar update method moves backwards as assets are added to the
loader queue after progress is updated. Total assets in queue increases
faster than they are loaded.

This is a dirty fix to only increase the progress bar if the percentage
increases. This will add inaccuracy (slow phase) to the lower end of the
bar but will fix any backwards movement.